feat: created a footer const file
updated branch
